+++ firmware/controllers/algo/defaults/default_base_engine.cpp
@@ -75,7 +75,7 @@ void setDefaultBaseEngine() {
 	engineConfiguration->primingDelay = 0.5;
 	engineConfiguration->vvtControlMinRpm = 500.0;
 
-    engineConfiguration->camDecoder2jzPosition = 75;
+    engineConfiguration->camDecoder2jzPosition = 95;
     engineConfiguration->camDecoder2jzPrecision = 25;
 
 	// Limits and Fallbacks